requirements.txt
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.0
schedule==1.2.1
